refactor: move package orderTable
