+++ src/main/java/kitchenpos/TableDao.java
@@ -1,0 +1,9 @@
+package kitchenpos;
+
+import kitchenpos.order.domain.repository.OrderTableDao;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TableDao extends OrderTableDao {
+
+}
