+++ src/main/java/kitchenpos/application/dto/TableGroupRequest.java
@@ -1,6 +1,7 @@
 package kitchenpos.application.dto;
 
 import java.util.List;
+import kitchenpos.order.application.dto.OrderTableRequest;
 
 public class TableGroupRequest {
     private List<OrderTableRequest> orderTables;
