+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -3,6 +3,7 @@ package kitchenpos.domain;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
+import kitchenpos.order.domain.OrderTable;
 import org.springframework.util.CollectionUtils;
 
 public class TableGroup {
