+++ src/main/java/kitchenpos/order/domain/repository/OrderTableDao.java
@@ -1,8 +1,8 @@
-package kitchenpos.dao;
+package kitchenpos.order.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 
 public interface OrderTableDao {
     OrderTable save(OrderTable entity);
