+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -7,15 +7,15 @@ import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
-import kitchenpos.application.dto.OrderTableRequest;
+import kitchenpos.order.application.dto.OrderTableRequest;
 import kitchenpos.application.dto.TableGroupRequest;
 import kitchenpos.application.dto.TableGroupResponse;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.order.domain.repository.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.BeforeEach;
