+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -4,15 +4,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalDateTime;
-import kitchenpos.application.dto.OrderTableRequest;
-import kitchenpos.application.dto.OrderTableRequest.NumberOfGuest;
-import kitchenpos.application.dto.OrderTableResponse;
+import kitchenpos.order.application.TableService;
+import kitchenpos.order.application.dto.OrderTableRequest;
+import kitchenpos.order.application.dto.OrderTableRequest.NumberOfGuest;
+import kitchenpos.order.application.dto.OrderTableResponse;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.dao.OrderTableDao;
+import kitchenpos.order.domain.repository.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
